Split into scheduler modules
